% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustifyRTMB.R
\name{logdet}
\alias{logdet}
\title{Compute the log-determinant of a matrix}
\usage{
logdet(x)
}
\arguments{
\item{m}{A square matrix}
}
\value{
The log-determinant of m
}
\description{
Compute the log-determinant of a matrix
}
