% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustifyRTMB.R
\name{robustify}
\alias{robustify}
\alias{robust_weight}
\title{Robustify likelihood components}
\usage{
robustify(x, robustness, robust_function = "id")

robust_weight(x, robustness, robust_function = "id")
}
\arguments{
\item{x}{A likelihood value}

\item{robustness}{A positive number. Larger numbers downweight components more.}

\item{robust_function}{Either "id" for the identity (non-robust), "ll" for log-logistic, or "ssh" for smooth semi-huber.}
}
\value{
A numeric value giving the robustified likelihood component.

A numeric value giving the robust weight.
}
\description{
Compute the robustified likelihood value
}
\section{Functions}{
\itemize{
\item \code{robust_weight()}: Compute the robust weight of a likelihood component

}}
