% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustifyRTMB-package.R
\docType{package}
\name{robustifyRTMB-package}
\alias{robustifyRTMB}
\alias{robustifyRTMB-package}
\title{robustifyRTMB: Robustify likelihood components in RTMB models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ethan Lawler \email{lawlerem@gmail.com} (\href{https://orcid.org/0000-0002-6225-3009}{ORCID})

}
\keyword{internal}
