% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_optimization.R
\name{robustly_optimize}
\alias{robustly_optimize}
\title{Fit a random effects model with a robustified likelihood}
\usage{
robustly_optimize(
  func,
  parameters,
  random,
  smooth,
  nodes,
  robust_schedule = 0,
  map = list(),
  bandwidth = 1e-06,
  silent = TRUE
)
}
\arguments{
\item{func}{See RTMB::MakeADFun}

\item{parameters}{See RTMB::MakeADFun}

\item{random}{See RTMB::MakeADFun}

\item{smooth}{A character naming the random effects that should be smoothed.}

\item{nodes}{A matrix giving the locations of the spline nodes}

\item{robust_schedule}{The sequence of robustness values to iterate through}

\item{map}{See RTMB::MakeADFun}

\item{bandwidth}{The bandwidth to use when smoothing the spline between iterations}

\item{silent}{Should the optimization tracing be suppressed?}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list containing:
}\if{html}{\out{</div>}}
\itemize{
\item par
The estimated parameter list
\item obj
The objective function from RTMB::MakeADFun
\item opt
The optimizer output from nlminb
\item sdr
The standard errors from RTMB::sdreport
}
}
\description{
Fit a random effects model with a robustified likelihood
}
